+++ jdaviz/app.py
@@ -475,16 +475,27 @@ class Application(VuetifyTemplate, HubListener):
     def load_data(self, file_obj, parser_reference=None, **kwargs):
         """
         Provided a path to a data file, open and parse the data into the
-        `~glue.core.data_collection.DataCollection` for this session. This also attempts to
-        find WCS links that exist between data components. Extra key word
-        arguments are passed to the parsing functions.
+        `~glue.core.data_collection.DataCollection` for this session.
+
+        For some parsers, this also attempts to find WCS links that exist
+        between data components.
 
         Parameters
         ----------
         file_obj : str or file-like
             File object for the data to be loaded.
+
+        parser_reference : str or `None`
+            The actual data parser to use. It must already be registered
+            to glue's data parser registry. This is mainly for internal use.
+
+        **kwargs : dict
+            Additional keywords to be passed into the parser defined by
+            ``parser_reference``.
+
         """
         self.loading = True
+
         try:
             try:
                 # Properly form path and check if a valid file
@@ -508,6 +519,7 @@ class Application(VuetifyTemplate, HubListener):
             data = self.state.settings.get('data', None)
             if parser_reference:
                 parser = data_parser_registry.members.get(parser_reference)
+
             elif data and isinstance(data, dict):
                 data_parser = data.get('parser', None)
                 if data_parser: